% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changemirror.R
\name{changemirror}
\alias{changemirror}
\title{This function just used for changing CRAN mirror
created by yujijun in 2021.01.26}
\usage{
changemirror()
}
\value{

}
\description{
This function just used for changing CRAN mirror
created by yujijun in 2021.01.26
}
